Fix problem where a fast client can "drown" a server
When a `yamux::Connection` is setting up a new incoming connection (`process_incoming()`) it tries to make as much progress as possible. If a client connection is fast enough to send all its data and then closes the connection, the `Connection` reads all the data within a single call to `process_incoming()`, including the end of the stream. This means that it will set `is_dead` to true on the connection which will cause `poll()` to return `Ready(None)` which bubbles up to the muxer where it is interpreted as if the connection was gone. The result is that `muxing::inbound_from_ref_and_wrap()` returns `None` and the data read is lost.
This PR re-orders thins such that `poll()` first checks for incoming connections and then check for a dead connection and clears the stream if needed.