Fix handling of receive window.
Except for `open_stream` always create a `StreamEntry` with a receive
window = DEFAULT_CREDIT. The remote has to assume we use the default. If
we configured a receive window size greater than that and use that
right away we will wait for more data while the remote is waiting for
for our window update. The end result is a deadlock.

This commit instantiates a `StreamEntry` in `on_data` and
`on_window_update` with a default receive window, matching the
expectation from remote and only when sending window updates will we use
the configured receive window size. The simple rule is: only change the
receive window size after telling the other end about it.