Limit pending buffer and fix test.
The cause for the sometimes not terminating test was that both endpoints
attempt to finish a writing operation and would not continue reading
from the socket. We must--as much as possible--continue reading, even
when we can not complete a write operation to prevent deadlock.

Since we buffer pending frames, which are meant to be delivered to the
remote and we make progress even if we can not write, we must limit the
resource growth of the pending frames buffer. This PR adds a new setting
`max_pending_frames` to do just that.