No RST for unknown streams if frame was FIN.
During shutdown, an endpoint may send a `FIN` for a stream which is
already dropped on the other end. Instead of sending back a `RST` we
ignore this frame because the sender will not send any more data for
this stream and if the stream was ever open, we have already sent a
`FIN` and `RST` frame for it.