Don't miss notifications.
If `poll_stream_notify`, `start_send_notify` or `poll_flush_notify` on a `Spawn`
are called prior to registering the current task with `self.tasks`, a notification
may be missed. I.e. `self.tasks` may be notified *before* the `NotReady` result
is returned, making it unsafe to insert the current task only upon receiving
`NotReady` from these functions.

This was observed in stalling benchmarks with concurrent streams. In particular,
if these benchmarks stalled, they always stalled on receiving the last frame of the last
stream, because the task notification was missed. As long as there is more traffic,
this problem is usually "papered over" by subsequent task wakups for further data, but
on the last frame, there are no more notifications as there is no more incoming data
on the connection, hence *if* the benchmarks stalled, it was always on the last frame
of the last stream.

This PR remedies the problem by ensuring that the current task is always registered
with `self.tasks` before calling these functions. This may result in some redundant
notifications but avoids missing any.