Rename `Handle::close` to `Handle::send_close`.
`Handle::send_close` will return a future that resolves as soon as the
close message has been sent to the connection actor.
`Handle::close` will instead wait for the connection to acknowledge that
it has marked all streams as closed. This helps clients which rely one
certain stream states in their own processing.