Change `Connection` error type to `Void`.
The connection actor has no supervisor, so we make sure it does not
itself produce an error. Instead we log fatal errors and transition to
`Done` state which is equivalent to aborting the connection.

The only error we have to consider in fact is that the `Sender` has
terminated and theoretically we may encounter this whenever we try to
send it a message and before we are informed about this in our role as
the supervisor of `Sender`.