Add `connected` flag to stream.
The current `read_after_close` handling within `Stream` is not what
`master` does. Instead of disallowing reads if the *connection* is
closed and `read_after_close` = false, it checks if the *stream* can
no longer read. To fix this, each stream needs to know when the
connection has been closed, hence we use the separate flag `connected`
to track this.