Refactor closing of connections.
The problem we face is that we have an unbounded number of `Control`s
which can all issue close commands. We want them all to be informed when
we have done so but we need to make sure that by closing the command
queue we do not generate false positives. The strategy here is to split
the command queue into two, one for streams and one for controls. As
soon as we receive a close (control) command, we close the stream
commands receiver, and only process any pending stream commands while
not processing any more control commands. Only after completing this
process do we resume processing control commands but only to inform all
controls that have issued a command that we are closed.