Make `Stream::state` private to the crate.
Keep the public API surface minimal. We may at some point replace the
`Mutex` around `Shared` with a futures-aware lock at which point
`Stream::state` would no longer be a synchronous method. As the method
is not used externally yet it is best not to introduce at this point.
Since we nevertheless use this in our tests, smoke.rs has been moved
crate.