Update to bytes-0.5
Some notable changes:

- `BytesMut::bytes_mut` now returns a `&mut [MaybeUninit<u8>]` as it may
contain uninitialised memory. Since we resize our buffer which implies
initialisation, we can just cast this slice to a `&mut [u8]` and use it
directly.

- `BytesMut::remaining_mut` no longer gives the length of the
`BytesMut::bytes_mut` as the `BufMut` implementation grows the buffer as
needed. So, to figure out if we need to reserve another BLOCKSIZE chunk we
now check how much capacity is left.

- `futures_codec` has been removed from tests as it has not updated to
bytes-0.5. The tests now work directly with the stream as `AsyncRead`
and `AsyncWrite` impl.