Map connection resets to `ConnectionError::Closed`.
When encountering an I/O error with `ErrorKind::ConnectionReset`, map
this to a `ConnectionError::Closed` value. The remote may have closed
the connection and a host implementing a "half-duplex" close sequence
would send us a RST when we send data because we have not noticed the
close yet. By mapping this case to a `ConnectionError::Closed` error,
we effectively treat this case like any other connection close.