Send back ACKs and use more principled casts.
Up to now we never replied to opened inbound streams with an ACK flag,
despite the fact that the spec clearly demands this behaviour. While we
continue to ignore any ACK flags we receive, with this change we properly
set an ACK flag if required.

Additionally, this PR makes casting of header and frame types more
principled. Instead of arbitrary casts only casts from `()` to `Data`,
`WindowUpdate` and `Ping` are allowed. We also model a binary sum type and
provide introductions from `T` to `Either<T, U>` and `T` to `Either<U, T>`
but no projections. This allows a more type-safe `StreamCommand` whose
`SendFrame` constructor only accepts `Frame<Either<Data, WindowUpdate>>`
values which is what we expect coming from streams.