Add an option to open outbound streams lazily.
The current behaviour is to always immediately send a `WindowUpdate`
frame to the remote when opening an outbound frame. While we do not wait
for an acknowledgement before we continue to send data over the stream,
we could also hold off and set the initial SYN flag in the first data
frame that is being sent over the new stream to the remote. The newly
added option `lazy_open` enables this. This should eventually become the
new default.

Implementation notice:

We currently check if we need to set flags when receiving a command from
the stream to send a frame. We might move this into the `stream` module
and make this decision local to the stream instead of looking up the
stream by ID. This would only require a bit more investigation for the
close case as the stream does currently not send a frame in this case on
which it could set the flag.