Fix race on sending of GC frames.
When a stream is dropped after sending some data, the
corresponding `SendFrame` commands may still be in the
queue when stream garbage collection kicks in. Stream
garbage collection conditionally sends final FIN or
RST frames, but it does so directly on the socket, thus
these frames may be sent (and hence received) before
other frames that may have been sent before the stream
was dropped. This is addressed here by sending the
final FIN or RST frames through the `stream_sender`,
so they are guaranteed to be sent last.

As a side-effect, these frames being sent also turn
up in the trace logs, which was not the case before.