Fix read-after-close.
Currently the `Stream` and `AsyncRead` impls for streams
check the condition for read-after-close based on whether
the stream command sender is closed (i.e. the receiver has
been dropped). However, after a substream has been closed,
dropping the receiver happens asynchronously as the connection
consumes the `Close` stream commands. As a result, it is
effectively possible within a short time window to do a
read-after-close even if this is disabled in the config.

This is addressed here by having the stream check the
shared state, which is the source of truth for the stream
state.