In 14a814e we missed updating the window.
While we reset the receive window in `Stream::send_pending` we forgot
to do so in case the initial send operation was successful. Because of
that we potentially issued multiple window updates to the remote which
would then have too much credit and their subsequent frames might even
be larger than the max. stream buffer size.

To fix this we now only use `Stream::send_pending` to send window
updates, rather than modifying the window in two more places.