Consume and return `self` in `next_stream`.
This ensures that futures are not cancelled and that the previous
call finished successfully.

This PR also returns `Stream` instead of `Option<Stream>` and signals
EOF via `ConnectionError::Closed` instead of `None`. This is motivated
by the fact that errors already mean the connection should be
considered closed and it is now impossible even to continue calling
`next_stream`.