src/tests: Write and read concurrently
Most of the tests in `src/tests.rs` have the structure of a client
sending data to a server which in turn echoes the data back to the
client which reads it in full and compares it to what it initially sent.

Some of the tests use randomly generated configurations where the
`WindowUpdateMode` can be either `OnRead` or `OnReceive`. When using
`WindowUpdateMode::OnRead` the client can not first send all data to the
server and only then receive all data from the server. In such case the
server would eventually exhaust its window to the client, thus not being
able to forward bytes to the client, thus not accepting new bytes from
the client, thus the client would be blocked on sending and thus the
whole test would deadlock.

With this commit the client writes and reads concurrently and thus does
not run into a deadlock when executing with `WindowUpdateMode::OnRead`.